+++ app/frontend/src/chatbots/publishone/components/Answer/Answer.tsx
@@ -1,11 +1,10 @@
-import publishoneLogo from "../../../../assets/publishone-chat.png";
+import publishoneLogo from "../../../../assets/publishone_logo.jpeg";
 import { createBotAnswer } from "../../../shared/answer";
 import { SpeechOutputBrowser } from "./SpeechOutputBrowser";
 import { SpeechOutputAzure } from "./SpeechOutputAzure";
-import styles from "./Answer.module.css";
 
 export const Answer = createBotAnswer(publishoneLogo, SpeechOutputBrowser, SpeechOutputAzure, {
     showAssistantName: false,
-    assistantLogoVariant: "wordmark",
-    assistantLogoClassName: styles.wordmarkLogo
+    showCopyButton: false,
+    assistantLogoPlacement: "outside-left"
 });